requirements.txt
flask
flask-login 
openai
langchain
sqlalchemy 
pyodbc
python-dotenv